영화 찜하기 controller 최종 수정

+++ popcornLovers/src/main/java/com/java4/popcorn/mypage/MypageController.java
@@ -13,6 +13,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.java4.popcorn.member.MemberDAO;
@@ -187,43 +188,85 @@ public class MypageController {
 	
 	
 	
+//	// 찜 추가 하기(빈 하트 >> 꽉찬 하트)
+//	@RequestMapping("mypage/addmoviejjim")
+//	public void addMovieJjim(MypageVO vo, HttpSession session) {
+//	    String member_id = (String) session.getAttribute("member_id"); // 로그인한 회원 세션
+//	    String movieId = vo.getMovieId(); // 선택한 movieId 값
+//
+//	    MypageVO movieInfo = dao.movieSelectOne(movieId); // dao에서 movie id값 읽어오기
+//
+//	    vo.setMovieId(movieInfo.getMovieId());
+//	    vo.setMember_id(member_id);
+//	    
+//	    if (member_id != null) { // 로그인이 되었을 때
+//	        int result = dao.addMovieJjim(vo); // 찜 추가 실행
+//	        if (result > 0) { // 찜 추가가 성공되면
+//	            session.setAttribute("jjimStatus_" + movieId, "liked"); // 세션에 찜한 영화 상태 정보 저장
+//	            System.out.println("추가 성공" + result);
+//	        } else { // 찜 추가가 실패
+//	            System.out.println("추가 실패" + result);
+//	        }
+//	    } else {
+//	        // 로그인이 되지 않았을 때
+//	        System.out.println("로그인 여부 확인");
+//	    }
+//	}
+	
+	
 	// 찜 추가 하기(빈 하트 >> 꽉찬 하트)
 	@RequestMapping("mypage/addmoviejjim")
 	public void addMovieJjim(MypageVO vo, HttpSession session) {
 	    String member_id = (String) session.getAttribute("member_id"); // 로그인한 회원 세션
 	    String movieId = vo.getMovieId(); // 선택한 movieId 값
 
-	    MypageVO movieInfo = dao.movieSelectOne(movieId); // dao에서 movie id값 읽어오기
-
-	    vo.setMovieId(movieInfo.getMovieId());
-	    vo.setMember_id(member_id);
-	    
 	    if (member_id != null) { // 로그인이 되었을 때
-	        int result = dao.addMovieJjim(vo); // 찜 추가 실행
-	        if (result > 0) { // 찜 추가가 성공되면
-	            session.setAttribute("jjimStatus_" + movieId, "liked"); // 세션에 찜한 영화 상태 정보 저장
-	            System.out.println("추가 성공" + result);
-	        } else { // 찜 추가가 실패
-	            System.out.println("추가 실패" + result);
+	        MypageVO movieInfo = dao.movieSelectOne(movieId); // dao에서 movie id값 읽어오기
+	        
+	        if (movieInfo != null) { // 찜 상태가 db에 없을 경우
+	            int result = dao.addMovieJjim(vo); // 찜 추가 실행 (찜 상태를 db에 저장)
+	            
+	            if (result > 0) { // 찜 추가가 성공되면
+	                session.removeAttribute("jjimStatus_" + movieId); // 세션에서 찜한 영화 상태 정보 제거
+	                System.out.println("추가 성공" + result);
+	            } else { // 찜 추가가 실패
+	                System.out.println("추가 실패" + result);
+	            }
+	        } else {
+	            System.out.println("영화 정보를 찾을 수 없음");
 	        }
 	    } else {
 	        // 로그인이 되지 않았을 때
 	        System.out.println("로그인 여부 확인");
 	    }
 	}
-
 	
 	// 찜 삭제하기(꽉찬 하트 >> 빈 하트)
 	@RequestMapping("mypage/removemoviejjim")
 	public void removeMovieJjim(MypageVO vo, HttpSession session) {
-		String member_id = (String) session.getAttribute("member_id");//로그인한 회원 세션
+	    String member_id = (String) session.getAttribute("member_id");//로그인한 회원 세션
 	    String movieId = vo.getMovieId(); // 선택한 movieId 값
 	    
-	    vo.setMember_id(member_id);
-	    dao.removeMovieJjim(vo);
-	    session.removeAttribute("jjimStatus_" + movieId);
+	    if (member_id != null) { // 로그인이 되었을 때
+	        dao.removeMovieJjim(vo);
+	        session.removeAttribute("jjimStatus_" + movieId); // 세션에서 찜한 영화 상태 정보 제거
+	    } else { // 로그인 되지 않았다면
+	        System.out.println("로그인 여부 확인");
+	    }
 	}
 
+	//찜 체크 확인하기
+	@RequestMapping("mypage/jjimCheck")
+	@ResponseBody
+	public int JjimCheck(MypageVO vo,  HttpSession session) {
+		String member_id = (String) session.getAttribute("member_id");//로그인한 회원 세션
+		vo.setMember_id(member_id);
+		
+		int jjimStatus = dao.jjim_Check(vo); //찜 상태 확인해서 jjimStatus에 넣어주기
+		return jjimStatus;
+	}
+	
+	
 	//마이페이지로 이동하기
 //	@RequestMapping("mypage/mypageGO")
 //	public String mypage(Model model) {
